Move `removeDeleteColumnMarker` to `columns`
